+++ src/main.rs
@@ -1,10 +1,13 @@
-use hs_mgt_game::cli::{describe_cli_error, run};
+use hs_mgt_game::cli::{describe_cli_error, eprint_error, run};
 
 fn main() {
   match run() {
     Ok(()) => {}
     Err(error) => {
-      eprintln!("Unable to run demo: {}", describe_cli_error(&error));
+      eprint_error(&format!(
+        "Unable to run demo: {}",
+        describe_cli_error(&error)
+      ));
       std::process::exit(1);
     }
   }